requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
dj-database-url==0.5.0
Django==3.2.8
django-ckeditor==6.4.2
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-js-asset==2.0.0
gunicorn==20.1.0
idna==3.3
mysqlclient==2.1.1
Pillow==9.2.0
psycopg2==2.9.3
psycopg2-binary==2.9.3
python-dateutil==1.5
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.11
whitenoise==6.2.0
